requirements.txt
annotated-types==0.7.0
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
cuid2==2.0.1
docutils==0.21.2
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.6.0
nh3==0.2.21
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyproject_hooks==1.2.0
python-ulid==3.0.0
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
twine==6.1.0
typing_extensions==4.13.0
urllib3==2.3.0
